Add an "advanced rating" mode for Hitman 2
This is based on Kotti's concept (https://rating-tool-23c0d.firebaseapp.com/)
and is still very much a WIP. The mode can be toggled from the Statman
right-click context menu.

Currently, the following events are tracked:

- Witness elimited (murder / accident)
- NPC pacified
- Caught trespassing
- Body found (dead / unconscious)
- Guards alerted
- Alarm triggered
- Guard / civilian killed
- Caught commiting crime

The alarm and caught commiting crime events are not tracked very
accurately right now.